build: move owner payout helpers into libbitcoin_common
The bloom filter and JSON writers in libbitcoin_common call GetOwnerPayouts
and PayoutListToJson, but these were defined in providertx.cpp which lives in
libbitcoin_node, so dash-tx (links common, not node) failed to link.

Move the layer-agnostic owner payout helpers (LegacyPayoutAsList,
GetOwnerPayouts, PayoutListToString, PayoutListToJson) into a new common
source file evo/providertx_util.cpp and add it to the dash-chainstate source
list. The trivial-validation helpers stay in providertx.cpp as node-only.
