+++ src/Makefile.am
@@ -931,6 +931,7 @@ libbitcoin_common_a_SOURCES = \
   deploymentinfo.cpp \
   evo/core_write.cpp \
   evo/netinfo.cpp \
+  evo/providertx_util.cpp \
   external_signer.cpp \
   governance/common.cpp \
   governance/core_write.cpp \
@@ -1200,6 +1201,7 @@ dash_chainstate_SOURCES = \
   evo/netinfo.cpp \
   evo/mnhftx.cpp \
   evo/providertx.cpp \
+  evo/providertx_util.cpp \
   evo/simplifiedmns.cpp \
   evo/smldiff.cpp \
   evo/specialtx.cpp \
